feat: half-hour-granular playtime — capture Steam/PSN sub-hour precision and clean up hours display (#645)
* docs: spec for issue #641 (half-hour playtime)

* docs: implementation plan for issue #641

* feat: half-hour-bucket recorded playtime display

* refactor: promote PlaytimeHours to float64 across adapter contract

* feat: capture Steam sub-hour playtime

* feat: preserve PSN sub-hour playtime to minute resolution

* test: guard fractional playtime through upsertPlatforms

* feat: accept half-hour values in manual playtime edit

* test: update game-edit-form total assertion for new format

* feat: apply half-hour bucket to HLTB times too

+++ internal/services/gog/library.go
@@ -15,7 +15,7 @@ type ExternalGameEntry struct {
 	ExternalID      string
 	Title           string
 	Platforms       []string // all platforms this product runs on
-	PlaytimeHours   int
+	PlaytimeHours   float64
 	OwnershipStatus string
 	IsSubscription  bool
 }
